requirements.txt
aiohttp==3.6.2
aniso8601==8.0.0
async-timeout==3.0.1
attrs==19.3.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
discord.py==1.3.3
Flask==1.1.2
Flask-RESTful==0.3.8
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.2
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.17.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.5
protobuf==3.11.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.14.0
SQLAlchemy==1.3.16
uritemplate==3.0.1
urllib3==1.25.9
websockets==8.1
Werkzeug==1.0.1
yarl==1.4.2
